requirements.txt
torch
torchvision
transformers==4.46.3
addict==2.4.0
yapf==0.43.0
timm==1.0.12
numpy==1.23.2
opencv-python
supervision==0.25.0
pycocotools
omegaconf==2.3.0
hydra-core==1.3.2
wandb==0.18.7